requirements.txt
streamlit
ultralytics
opencv-python
numpy
pillow
pandas